requirements.txt
asgiref==3.2.10
Django==3.1.1
django-jalali-date==0.3.1
et-xmlfile==1.0.1
executing==0.5.2
jdatetime==3.6.2
jdcal==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
varname==0.4.0
xlrd==1.2.0
